+++ CHANGELOG.md
@@ -42,6 +42,7 @@ All notable changes to eww will be listed here, starting at changes since versio
 - Add `value-pos` to scale widget (By: ipsvn)
 - Add `floor` and `ceil` function calls to simplexpr (By: wsbankenstein)
 - Add `formatbytes` function calls to simplexpr (By: topongo)
+- Add `onfocus` and `onfocuslost` properties to eventbox (By: RelativeAlbatros)
 
 ## [0.6.0] (21.04.2024)